honeypi beta
What’s new
Bug fixes
Hotfix for A9X devices, which could not boot on 14.5 at all with 0.12.3
Fix package dependency issues on the deb repo
Notes
A11 devices on iOS 14.0 and above require removing the passcode and enabling “Skip A11 BPR check” in the options. This is not recommended, but it’s there.
Known issues
A7 devices will currently not work with the Linux version
The exploit may not work as reliably on some devices, such as the Raspberry Pi Zero and Raspberry Pi 3
Linux GUI build when using CLI mode doesn’t support auto assertion of DFU mode on advanced Apple TV 4K breakout boards
bridgeOS:
If you disconnect the cable, then as soon as macOS boots it’ll take over the USB connection and disallow communication with the T2
Some users have reported checkra1n crashing when Chinese is set as the only system language. It is unknown whether any other language is affected.
Unsupported platforms
This beta is only available for macOS and Linux. Work is ongoing to support Windows, which will be added in a later release.

Package managers
At the moment, checkra1n only supports installing selio or zebra. Support for other package managers, including Zebra and Installer is coming soon and is expected within the coming weeks.